backend/requirements.txt
flask
flask-cors
psycopg2-binary

prometheus-client==0.20.0
